% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaCalc.R
\name{area_stats_seq}
\alias{area_stats_seq}
\title{Forest area calculation single mode (Helper Function)}
\usage{
area_stats_seq(studysite, inputForestMap, latlon)
}
\arguments{
\item{studysite}{An sf object}

\item{inputForestMap}{A raster object}

\item{latlon}{Logical indicating if raster is unprojected or not}
}
\value{
A dataframe with estimated areas
}
\description{
Forest area calculation single mode (Helper Function)
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
