% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaCalc.R
\name{AreaCalc}
\alias{AreaCalc}
\title{Area of forest change}
\usage{
AreaCalc(
  inputForestMap = NULL,
  studysite = NULL,
  latlon = NULL,
  polyName = NULL,
  ncores = 1,
  saveCSV = FALSE,
  years = 2001:2018
)
}
\arguments{
\item{inputForestMap}{A \code{RasterLayer} object with forest cover represented
in binary (e.g 0 represents no forest; 1 represents forest). Each layer is
expected to represent one year in the \code{years} object in a consecutive order.}

\item{studysite}{An object of type \code{sf} with a given
number of polygons defining the areas of interest. Forest area statistics
will be calculated for features in the \code{studysite} object.}

\item{latlon}{\code{logical}: Indicates whether or not the
\code{inputForestMap} object is based on a geographic coordinate system
or is projected to a planar coordinate system. In the former case,
area is approximated by \code{\link[raster:area]{raster::area}}.}

\item{polyName}{\code{charachter} of length 1. Indicates the column in the
data frame of the \code{studysite} object to uniquely identify the features
of interest. The function will fail if there is no unique identification of
the polygons}

\item{ncores}{The number of cores to use, i.e. at most how many child
processes will be run simultaneously.}

\item{saveCSV}{Default is \code{FALSE}. You can specify a directory on your
local machine where the results area saved in \code{.csv} format. The
features will be identified by the column specified in \code{polyName}.}

\item{years}{A vector of type \code{numeric} indicating the years which
are represented by pixels. For GFW data and the default these are 2001 to 2018.}
}
\value{
The \code{studysite} object with its data frame appended columnwise
 for every single entry in the \code{years} object representing the total area
 of forest in a given layer of the \code{inputForestMap} object.
 If \code{latlon=TRUE} the returned area is in \emph{km²}, otherwise in the
 squared unit of the input projection (most commonly in meters).
}
\description{
Calculates the changes in forest area based on a \code{rasterObject} forest
  mask and a \code{SpatialPolygonsDataFrame} containing areas of interest.
}
\note{
This function relies heavily on parallization, indicating the
  importance of both, a high number of CPUs and large enough RAM.
}
\examples{
library(sf)
library(raster)
library(mapme.forest)

aoi = st_read(system.file("extdata", "aoi_polys.gpkg", package = "mapme.forest"))
yearlyRaster = stack(system.file("extdata", "pkgTest_yearlyCover.tif",
                                 package = "mapme.forest"))
result = AreaCalc(inputForestMap = yearlyRaster,
                  studysite = aoi[1,],
                  latlon = TRUE,
                  polyName = "id",
                  ncores = 1,
                  saveCSV = FALSE,
                  years = 2000:2018)
str(result)

}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
