% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragStatsCalc.R
\name{calcFragStats_seq}
\alias{calcFragStats_seq}
\title{Helper function to calculate Fragstats for a single polygon in sequential mode}
\usage{
calcFragStats_seq(studysite, inputRasterFiles, FragStats, latlon)
}
\arguments{
\item{studysite}{A one-row sf-object}

\item{inputRasterFiles}{A Raster*-Object.}

\item{FragStats}{A charachter vector.}
}
\description{
Helper function to calculate Fragstats for a single polygon in sequential mode
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
