% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2Calc.R
\name{co2_calc_seq}
\alias{co2_calc_seq}
\title{CO2 emission calculation single mode (Helper Function)}
\usage{
co2_calc_seq(inputForestMap, inputLossMap, inputCO2Map, studysite, unis)
}
\arguments{
\item{inputForestMap}{A raster object}

\item{inputLossMap}{A raster layer object}

\item{inputCO2Map}{A raster layer object}

\item{studysite}{A sf object}

\item{unis}{A numeric vector indicating yearly cell values}
}
\value{
A data.frame object with Yearly sums of co2 emissions by tree cover loss
}
\description{
CO2 emission calculation single mode (Helper Function)
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
