% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coverratio}
\alias{coverratio}
\title{Helper Function to calculate cover ratio}
\usage{
coverratio(r, p)
}
\arguments{
\item{r}{A raster object}

\item{p}{A sf object}
}
\value{
The cover ratio in percent
}
\description{
This function is used inside of the area calculation functions to evaluate
the ratio a given raster is not covered by a polygon. Can be used as an indicator
if further cropping of the raster ca improve calculation time.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
