% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGFW.R
\name{downloadfGFW}
\alias{downloadfGFW}
\title{Download GFW and CO2-Emission data}
\usage{
downloadfGFW(
  shape,
  dataset = "GFC-2018-v1.6",
  basename = "Hansen_1.6",
  outdir = ".",
  keepTmpFiles = F,
  .tmpdir = "./.tmp/"
)
}
\arguments{
\item{shape}{A \code{sfObject} determining the extent for which to get the GFW data.}

\item{dataset}{A \code{charachter} specifiying the version of the GFW data to extract.
Defaults to version 1.6 from 2018. The newer version can be downloaded when
you specify \code{dataset = "GFW-2019-v1.7"}.}

\item{basename}{A \code{charachter} which will be added to the resulting
file names.}

\item{outdir}{A \code{charachter} for a local directory for the final outputs.
If it is not existing it will be created without a message. Defaults to the
current working directory.}

\item{keepTmpFiles}{A \code{logical} indicating if the raw tiles should be kept in
the \code{tmpdir} directory. It defaults to \code{FALSE}, which usually makes sense.
However, if you apply a number of analysis for different extents but there might
be overlaps, it could be useful to keep the downloaded files on your machine.
The function will only download the necessary data if it is not alread present
in the \code{tmpdir}.}

\item{.tmpdir}{A \code{charachter} indicating a directory where to download
the raw tiles to. It defaults to a hidden folder called \code{.tmp} in the
current working directory, but can be set by the user. If the directory does
not exist on your machine it will be created without a message.}
}
\value{
A \code{vector} of type \code{charachter} with all the files matching
 the \code{basename} pattern in the \code{outdir} directory.
}
\description{
This function downloads the necessary Global Forest Watch and CO2 Emission
 data for a given shapefile. The shapefile is transformed to its bounding
 box, projected to its central UTM Zone and a buffer is applied. The default
 buffer size is 5000 meteres, but it can be changes by the user. It uses
 gdal system calls to speed up the process. The raw files by default will be
 deleted, but they also can be kept on the disk. Note, the CO2 emission data
 is only available roughly between 30N and 30S and the function will fail if
 your input shapfiles extent plus buffer is extending beyond those limits.
 This function relies heavily on code published at https://github.com/azvoleff/gfcanalysis.
}
\note{
This function depends on available gdal binaries on your system. Make sure they
  are available as environment variables on your machine or use our docker image instead.
}
\examples{
\dontrun{
aoi = st_read(system.file("extdata", "aoi_polys.gpkg", package = "mapme.forest"))

raster_files = downloadfGFW(shape = aoi,
                            basename = "pkgTest",
                            outdir = "./data/",
                            keepTmpFiles = T,
                            .tmpdir = "./data/tmp")

# resulting rasters are not automatically cropped to the extent of aoi
rasters = stack(lapply(raster_files, function(f){
f = brick(f)
f = crop(f, aoi)
}))
}

}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
