% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{funTrendsDir}
\alias{funTrendsDir}
\title{Calculate Sens Slope Direction}
\usage{
funTrendsDir(y)
}
\arguments{
\item{y}{A numeric vector for which to calculate the sens slope.}
}
\value{
The direction of the sens slope.
}
\description{
This function calculates the sens slope direction at the 95% confidence interval
 in cases where more than 75% of the values ar not NA. Otherwise NA is returned.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
