% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{funTrendsSig}
\alias{funTrendsSig}
\title{Calculate Sens Slope Trend Significance}
\usage{
funTrendsSig(y)
}
\arguments{
\item{y}{A numeric vector for which to calculate the sens slope.}
}
\value{
The p-value of the sens slope.
}
\description{
This function calculates the sens slope significance at the 95% confidence interval
 in cases where more than 75% of the values are not NA. Otherwise NA is
 returned.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
