% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getTM}
\alias{getTM}
\title{Get yearly binary tree cover maps}
\usage{
getTM(inputForestMap, inputLossMap, years = 2000:2018)
}
\arguments{
\item{inputForestMap}{A \code{RasterLayer} with a binary mask of forest cover
(with 0 representing no forest; 1 representing forest) for the starting year.}

\item{inputLossMap}{A \code{RasterLayer} in which the cell values
represent the year in which forest loss occured during the time series
represented in the \code{years} object.}

\item{years}{A \code{numeric} vector indicating the years represented in the
values of the \code{inputLossMap} object. 0 is expected to indicate no loss
at all, a value of 1 corresponds to the first value in \code{years}, a value
of 2 to the second, and so on.}
}
\value{
A binary \code{RasterStack} with a number of layers equal to \code{length(years)}.
 A value of 0 represents no forest cover, a value of 1 represents forest cover.
 All cell values indicating forest loss in the \code{inputLossMap} object at a given year
 will be consectuivley set to 0. The result is a yearly binary classification of forest cover.
}
\description{
This function calculates yearly tree cover maps based on a starting year layer
in a binary forest mask form and an loss year layer which indicates which pixels
were subject to forest loss in a given year,
}
\examples{
library(raster)
library(mapme.forest)

binaryCover = raster(system.file("extdata", "pkgTest_binaryCover.tif",
                                 package = "mapme.forest"))
lossYear = raster(system.file("extdata", "pkgTest_lossyear.tif",
                              package = "mapme.forest"))
yearlyMaps = getTM(inputForestMap = binaryCover,
                   inputLossMap = lossYear,
                   years = 2001:2018)
yearlyMaps

}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
