% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossCalc.R
\name{LossCalc}
\alias{LossCalc}
\title{Area of Forest Loss}
\usage{
LossCalc(
  inputForestMap = NULL,
  inputLossMap = NULL,
  studysite = NULL,
  latlon = NULL,
  polyName = NULL,
  ncores = 1,
  saveCSV = FALSE,
  years = 2001:2018
)
}
\arguments{
\item{inputForestMap}{A \code{RasterLayer} object with forest cover represented
in binary (e.g 0 represents no forest; 1 represents forest). Each layer is
expected to represent one year in the \code{years} object in a consecutive order.}

\item{inputLossMap}{A \code{RasterLayer} object with values indicating the
year tree cover was lost. The annual area of tree cover lost is calculated
based on the values found in this data set.}

\item{studysite}{An object of type \code{sf} with a given
number of polygons defining the areas of interest. Forest area statistics
will be calculated for features in the \code{studysite} object.}

\item{latlon}{\code{logical}: Indicates whether or not the
\code{inputForestMap} object is based on a geographic coordinate system
or is projected to a planar coordinate system. In the former case,
area is approximated by \code{\link[raster:area]{raster::area}}.}

\item{polyName}{\code{charachter} of length 1. Indicates the column in the
data frame of the \code{studysite} object to uniquely identify the features
of interest. The function will fail if there is no unique identification of
the polygons}

\item{ncores}{The number of cores to use, i.e. at most how many child
processes will be run simultaneously.}

\item{saveCSV}{Default is \code{FALSE}. You can specify a directory on your
local machine where the results area saved in \code{.csv} format. The
features will be identified by the column specified in \code{polyName}.}

\item{years}{A vector of type \code{numeric} indicating the years which
are represented by pixels. For GFW data and the default these are 2001 to 2018.}
}
\value{
The studysite object with it's attribute table being ammended by the
  results of the calculation. Additionally, a csv file containing the \code{polyName}
  attribute as well as the results of the calculation can be saved to disk.
}
\description{
Calculation of the area where forest loss occured in a tree cover time series.
It uses information on the occurence of forest loss and a forest mask to focus
the analysis to pixles where lost occured. The area is then summed up for every
polygon specified in the studysite object.
}
\note{
This function relies heavily on parallization, indicating the
  importance of both, a high number of CPUs and large enough RAM.
}
\examples{
library(sf)
library(raster)
library(mapme.forest)

aoi = st_read(system.file("extdata", "aoi_polys.gpkg", package = "mapme.forest"))
yearlyRaster = stack(system.file("extdata", "pkgTest_yearlyCover.tif",
                                  package = "mapme.forest"))
lossYear = raster(system.file("extdata", "pkgTest_lossyear.tif",
                              package = "mapme.forest"))

result = LossCalc(inputForestMap = yearlyRaster,
                 inputLossMap = lossYear,
                 studysite = aoi[1,],
                 latlon = TRUE,
                 polyName = "id",
                 ncores = 2,
                 saveCSV = FALSE,
                 years = 2000:2018)
str(result)

}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
