% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGFW.R
\name{makeGFWGrid}
\alias{makeGFWGrid}
\title{Create the GFW tile grid}
\usage{
makeGFWGrid(mnx, mxx, dx, mny, mxy, dy, proj = NULL)
}
\arguments{
\item{mnx}{Minimum x corrdinate}

\item{mxx}{Maximum x coordinate}

\item{dx}{x resoltion}

\item{mny}{Minimum y coordinated}

\item{mxy}{Maximum y coordinated}

\item{dy}{y resolution}

\item{proj}{projection as EPSG code}
}
\value{
A grid.
}
\description{
This function is used in \code{\link{downloadfGFW}} to create a grid representing
the GFW tiles.
}
\keyword{internal}
