% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassStat2.R
\name{shape.index}
\alias{shape.index}
\title{Shape index}
\usage{
shape.index(a, p)
}
\arguments{
\item{a}{A}

\item{p}{P}
}
\value{
Shape index value
}
\description{
Helper function for the calculation of the shape in \code{\link{ClassStat2}}.
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
\keyword{internal}
