% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.R
\name{statsGRASS}
\alias{statsGRASS}
\title{Translation pipeline using GRASS}
\usage{
statsGRASS(
  grass,
  addon_base,
  areas,
  tree_cover,
  tree_loss,
  tree_co2,
  idcol,
  thresholdClump,
  thresholdCover,
  years,
  outdir = NULL,
  saveRaster
)
}
\arguments{
\item{grass}{A character vector pointing to the directory where GRASS binaries are installed.}

\item{addon_base}{A character vector pointing to the directory where GRASS add-on binaries are installed.
In order to work, the (\code{r.area})[https://grass.osgeo.org/grass78/manuals/addons/r.area.html]
add-on needs to be installed. See (here)[https://grass.osgeo.org/download/addons/] to learn
how to install GRASS add-ons.}

\item{areas}{An \code{sf} object with polygons representing areas of interest for which to
preprocess individual rasters.}

\item{tree_cover}{A character vector pointing to the raster file representing
tree cover in percentage for the entire study domain.}

\item{tree_loss}{A character vector pointing to the raster file representing
the year of tree cover loss for the entire study domain.}

\item{tree_co2}{A character vector pointing to the raster file representing
co2 emissions from tree cover loss for the entire study domain.}

\item{idcol}{A character vector identifying the a column name which uniquely
identifies the polygons in the \code{areas} object. Values  will be used to name
the output rasters.}

\item{thresholdClump}{A numeric value identifying the number of pixels of smallest clumps.
All raster cell clumps smaller than the specified threshold will be removed.}

\item{thresholdCover}{A numeric value identifying the smallest cover percentage
to be considered as forest. All raster cell values below this threshold will be
removed}

\item{years}{A numeric vector specifying for which years to calculate an annual
forest mask starting from 2001. (e.g. years = c(2001:2010)).}

\item{outdir}{A character pointing to a directory where the output files
will be written to. If files with the same name as specified by the values in
\code{idcol} are present, their calculation is skipped without warning.}

\item{saveRaster}{Logical indicating if raster are to be saved to disk}
}
\description{
Translation pipeline using GRASS
}
\author{
Darius Görgen, Dr. Fabian Löw (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} KfW FZ Evaluierung \email{FZ-Evaluierung@kfw.de}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} KfW Bankengruppe
\cr
\emph{License:} GPL-3
}
