% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassStat2.R
\name{aggregation.index}
\alias{aggregation.index}
\title{Aggregation index}
\usage{
aggregation.index(a, g)
}
\arguments{
\item{a}{A}

\item{g}{G}
}
\value{
Aggregation index in percent.
}
\description{
Helper function for the calculation of the aggregation index in \code{\link{ClassStat2}}.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
