% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaCalc.R
\name{area_stats_seq}
\alias{area_stats_seq}
\title{Forest area calculation single mode (Helper Function)}
\usage{
area_stats_seq(studysite, inputForestMap, latlon)
}
\arguments{
\item{studysite}{An sf object}

\item{inputForestMap}{A SpatRaster object}

\item{latlon}{Logical indicating if raster is unprojected or not}
}
\value{
A dataframe with estimated areas
}
\description{
Forest area calculation single mode (Helper Function)
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net} Johannes Schielein
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} \email{contact@mapme-initiative.org}
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
