% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepTC}
\alias{prepTC}
\title{Pre-Processing of tree cover layers}
\usage{
prepTC(
  inputForestMap,
  thresholdCover = NULL,
  thresholdClump = NULL,
  ncores = 1
)
}
\arguments{
\item{inputForestMap}{A RasterLayer representing forest cover.
For the primary use-case of this function tree cover is expected
as values in the range [0,100] representing percentage and is used
when the user wants to obtain a binary tree cover map.}

\item{thresholdCover}{A numeric vector of length one in the range of
[0,100]. It is used as the definition criterium for forest areas in
datasets like Hansen (2013), where a pixel's values indicates the
percentage of coverage with trees. Any pixel below this threshold
will get the value 0, any pixel above will get the value 1.}

\item{thresholdClump}{A numeric of length one indicating the number of
pixels which are classified as a clump. Any number of pixels smaller than
this threshold will be classified to NA.}

\item{ncores}{Number of cores to be used in internal calcuation functions. Defaults to 1 core but multiple can be used.}
}
\value{
A RasterLayer object with binary values when \code{thresholdCover} was
  specified and with raster cell clumps removed in case \code{thresholdClump} was
  specified. Either one of the parameters can be specified on its own.
}
\description{
Calculates a binary tree cover mask for a baseline year.
}
\examples{
library(terra)
library(mapme.forest)

treeCover = rast(system.file("extdata", "pkgTest_treecover2000.tif",
                               package = "mapme.forest"))
binaryCover = prepTC(inputForestMap = treeCover,
                             thresholdCover = 75,
                             thresholdClump = 25)
binaryCover

}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}, Johannes Schielein
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact:}  \email{contact@mapme-initiative.org}
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
