% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassStat2.R
\name{shape.index}
\alias{shape.index}
\title{Shape index}
\usage{
shape.index(a, p)
}
\arguments{
\item{a}{A}

\item{p}{P}
}
\value{
Shape index value
}
\description{
Helper function for the calculation of the shape in \code{\link{ClassStat2}}.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
\keyword{internal}
